+++ lib/exop_data/generators/float.ex
@@ -3,12 +3,14 @@ defmodule ExopData.Generators.Float do
   Implements ExopData generators behaviour for `float` parameter type.
   """
 
+  alias ExopData.Generators.AliasResolvers.Numericality
+
   @behaviour ExopData.Generator
 
   @diff 0.1
 
   def generate(opts \\ %{}, _props_opts \\ %{}) do
-    opts |> Map.get(:numericality) |> do_generate()
+    opts |> Map.get(:numericality) |> Numericality.resolve_aliases() |> do_generate()
   end
 
   @spec do_generate(map()) :: StreamData.t()
@@ -18,19 +20,49 @@ defmodule ExopData.Generators.Float do
 
   defp do_generate(%{is: exact}), do: constant(exact)
 
-  defp do_generate(%{min: min} = contract) do
-    contract
-    |> Map.delete(:min)
-    |> Map.put(:greater_than_or_equal_to, min)
-    |> do_generate()
-  end
+  defp do_generate(%{eq: exact}), do: constant(exact)
 
-  defp do_generate(%{max: max} = contract) do
-    contract
-    |> Map.delete(:max)
-    |> Map.put(:less_than_or_equal_to, max)
-    |> do_generate()
-  end
+  # defp do_generate(%{gt: gt} = contract) do
+  #   contract
+  #   |> Map.delete(:gt)
+  #   |> Map.put(:greater_than, gt)
+  #   |> do_generate()
+  # end
+
+  # defp do_generate(%{gte: gte} = contract) do
+  #   contract
+  #   |> Map.delete(:gte)
+  #   |> Map.put(:greater_than_or_equal_to, gte)
+  #   |> do_generate()
+  # end
+
+  # defp do_generate(%{min: min} = contract) do
+  #   contract
+  #   |> Map.delete(:min)
+  #   |> Map.put(:greater_than_or_equal_to, min)
+  #   |> do_generate()
+  # end
+
+  # defp do_generate(%{lt: lt} = contract) do
+  #   contract
+  #   |> Map.delete(:lt)
+  #   |> Map.put(:less_than, lt)
+  #   |> do_generate()
+  # end
+
+  # defp do_generate(%{lte: lte} = contract) do
+  #   contract
+  #   |> Map.delete(:lte)
+  #   |> Map.put(:less_than_or_equal_to, lte)
+  #   |> do_generate()
+  # end
+
+  # defp do_generate(%{max: max} = contract) do
+  #   contract
+  #   |> Map.delete(:max)
+  #   |> Map.put(:less_than_or_equal_to, max)
+  #   |> do_generate()
+  # end
 
   defp do_generate(%{greater_than: greater_than, less_than: less_than}) do
     StreamData.filter(