+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - formatter rules
 - ex_doc 0.20
 - elixir >= 1.6.0
-- :inner check accepts Keyword as well as Map
+- `:inner` check accepts Keyword as well as Map
+- `:numericality` aliases support (see Exop docs)
 
 ## [0.1.6] - 2019.03.29